[zengzm][teei]微可信p5 patch,打开安全校验 [(CR)]
Reviewed-on: http://sc.lenovo.com:8080/422833
Tested-by: 廖沪俊 <liaohj@lenovo.com>

+++ drivers/misc/mediatek/teei/V1.0/tz_driver/include/global_function.h
@@ -4,7 +4,7 @@ extern int forward_call_flag;
 extern int irq_call_flag;
 extern int fp_call_flag;
 extern int keymaster_call_flag;
-extern int teei_vfs_flag;
+extern unsigned long teei_vfs_flag;
 extern struct completion global_down_lock;
 extern unsigned long teei_config_flag;
