+++ drivers/misc/mediatek/teei/V1.0/tz_driver/include/irq_register.h
@@ -67,5 +67,5 @@ extern int fp_call_flag;
 extern int keymaster_call_flag;
 static struct work_entry work_ent;
 extern struct work_queue *secure_wq;
-extern int teei_vfs_flag;
+extern unsigned long teei_vfs_flag;
