+++ drivers/misc/mediatek/teei/V1.0/tz_driver/switch_queue.c
@@ -60,7 +60,7 @@ extern struct kthread_worker ut_fastcall_worker;
 extern int forward_call_flag;
 extern int fp_call_flag;
 extern int keymaster_call_flag;
-extern int teei_vfs_flag;
+extern unsigned long teei_vfs_flag;
 extern int irq_call_flag;
 
 extern void nt_sched_t_call(void);
